update sak_dir and sak_sys 	modified:   SConstruct 	deleted:    debug/program.c 	deleted:    debug/program.h 	renamed:    debug/log.c -> debug/sak_log.c 	renamed:    debug/trace.c -> debug/sak_trace.c 	deleted:    include/log.h 	new file:   include/sak.h 	new file:   include/sak/sak_dir.h 	new file:   include/sak/sak_log.h 	renamed:    include/macros.h -> include/sak/sak_macros.h 	new file:   include/sak/sak_sys.h 	new file:   include/sak/sak_trace.h 	deleted:    include/sak_dir.h 	deleted:    include/trace.h 	modified:   io/epool.c 	modified:   io/token_bucket.c 	modified:   mem/mem.c 	modified:   mem/mempool.c 	deleted:    platform/platform.h 	renamed:    platform/linux.c -> sys/linux.c 	renamed:    platform/sak_dir.c -> sys/sak_dir.c 	copied:     platform/windows.c -> sys/sak_sys.c 	renamed:    platform/windows.c -> sys/windows.c 	modified:   test/dir_test.c 	modified:   test/epool_test.c 	modified:   test/esignal_test.c 	modified:   test/macro_test.c 	modified:   test/mem_test.c 	modified:   test/mempool_test.c 	modified:   test/queue_test.c 	modified:   test/stack_test.c 	modified:   test/trace_test.c

+++ test/mempool_test.c
@@ -1,13 +1,13 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+#include "sak.h"
 
 #include "mempool.h"
 #include "mem.h"
-#include "log.h"
 
 #define BLKSIZE (1024)
-#define TESTSIZE (1024*1024*1024)
+#define TESTSIZE (1024*1024*1024ull)
 
 int main(int argc, char *argv[])
 {
@@ -24,7 +24,7 @@ int main(int argc, char *argv[])
             mem1 = mempool_get(mp);
         }
         uint64_t size = mempool_memsize(mp);
-        printf("------memsize: %llu, oobsize: %llu, oobrate: %f\n", size, size - TESTSIZE, (size - TESTSIZE)*1.0/TESTSIZE);
+        printf("------memsize: %llu, oobsize: %llu, oobrate: %f\n", (ull)size, (ull)size - TESTSIZE, (size - TESTSIZE)*1.0/TESTSIZE);
     }else{
         for (i = 0; i < TESTSIZE/BLKSIZE; i++){
             mem1 = malloc(BLKSIZE);
